chore: Implement a change in how project metadata is stored
This change enables sharing metadata across different layout files, ensuring consistent and easily modifiable metadata.
